move all routes into folders, need to put all the consts required
